# backend/app/models/acquisition_method.py
from sqlalchemy import Integer, String, Text
from sqlalchemy.orm import Mapped, mapped_column, relationship

from app.models.base import Base


class AcquisitionMethod(Base):
    __tablename__ = "acquisition_methods"

    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
    name: Mapped[str] = mapped_column(String(255), nullable=False, unique=True)
    description: Mapped[str | None] = mapped_column(Text)

    specimens = relationship("Specimen", back_populates="acquisition_method")